#21 SecurityConfig divided into SecurityConfig and AuthManagerConfig.
